Add Friend

+++ InterOn/InterOn.Api/Controllers/FriendController.cs
@@ -1,9 +1,12 @@
 ﻿using System.Threading.Tasks;
 using InterOn.Service.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace InterOn.Api.Controllers
-{   [Route("/api/user/{userId}/friends")]
+{
+    [Authorize]
+    [Route("/api/user/{userId}/friends")]
     public class FriendController : Controller
     {
         private readonly IUserFriendService _service;
@@ -16,7 +19,15 @@ namespace InterOn.Api.Controllers
         public async Task<IActionResult> AddFriend(int userId)
         {
             var userIdLogged = int.Parse(HttpContext.User.Identity.Name);
+            if (!await _service.IsExistUser(userId))
+            {
+                return BadRequest("Nie ma takiego użytkowika o tym Id");
+            }
 
+            if (await _service.IsExistFriendship(userIdLogged, userId))
+            {
+                return BadRequest("Już dodałeś tą osobę do znajomych");
+            }
             if (userIdLogged == userId)
             {
                 return BadRequest("Ten sam użytkownik nie może dodać się do znajomych ");
@@ -29,12 +40,19 @@ namespace InterOn.Api.Controllers
         public async Task<IActionResult> ConfirmFriend(int userId)
         {
             var userIdLogged = int.Parse(HttpContext.User.Identity.Name);
-
+            if (!await _service.IsExistUser(userId))
+            {
+                return BadRequest("Nie ma takiego użytkowika o tym Id");
+            }
             if (userIdLogged == userId)
             {
                 return BadRequest("Ten sam użytkownik nie może potwierdzić znajomych ");
             }
 
+            if (!await _service.IsExistFriendship(userIdLogged, userId))
+            {
+                return BadRequest("Nie ma takiej znajomości do zatwierdzienia");
+            }
             await _service.ConfirmFriend(userIdLogged, userId);
             return Ok();
         }
@@ -46,9 +64,38 @@ namespace InterOn.Api.Controllers
             {
                 return BadRequest("Ten sam użytkownik nie może usunąć siebie ze znajomych");
             }
-
-            await _service.RemoveFriendAsync(userIdLogged, userId);
+            if (!await _service.IsExistUser(userId))
+            {
+                return BadRequest("Nie ma takiego użytkowika o tym Id");
+            }
+            if (!await _service.IsExistFriendship(userIdLogged, userId))
+            {
+                return BadRequest("Nie ma takiech znajomych");
+            }
+             await _service.RemoveFriendAsync(userIdLogged, userId);
             return Ok();
+        }
+        [HttpGet]
+        public async Task<IActionResult> GetConfirmedFriends(int userId)
+        {
+            if (!await _service.IsExistUser(userId))
+            {
+                return BadRequest("Nie ma takiego użytkowika o tym Id");
+            }
+            var userIdLogged = int.Parse(HttpContext.User.Identity.Name);
+            var res = await _service.GetConfirmedFriendsAsync(userIdLogged);
+            return Ok(res);
+        }
+        [HttpGet("invitation")]
+        public async Task<IActionResult> GetInvitaionFriends(int userId)
+        {
+            if (!await _service.IsExistUser(userId))
+            {
+                return BadRequest("Nie ma takiego użytkowika o tym Id");
+            }
+            var userIdLogged = int.Parse(HttpContext.User.Identity.Name);
+            var res = await _service.GetInvFriendsAsync(userIdLogged);
+            return Ok(res);
         }
     }
 }
